app/requirements.txt
fastapi
supabase-py
python-dotenv
pydantic
sqlalchemy